pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "use-anything"
version = "0.1.0"
description = "Generate agent skills from existing software interfaces"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Use-Anything"}]
dependencies = [
  "click>=8.1.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.0",
  "tiktoken>=0.7.0",
  "rich>=13.7.0",
  "jsonschema>=4.23.0",
  "anthropic>=0.34.0",
  "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
]

[project.scripts]
use-anything = "use_anything.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
